% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpseR.R
\name{availabel_prop}
\alias{availabel_prop}
\title{get availabel property}
\usage{
availabel_prop(note = "stdprot")
}
\arguments{
\item{note}{Sequence definition types. There are four internal types.{stdprot|didna|dirna|tridna})}
}
\description{
get availabel property
}
\author{
Lang Zhou
}
