% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{available_prop}
\alias{available_prop}
\title{get available property}
\usage{
available_prop(note = "stdprot")
}
\arguments{
\item{note}{Sequence definition types. There are four internal types.{stdprot|didna|dirna|tridna})}
}
\description{
get available property
}
\examples{
 
available_prop(note = "didna")
available_prop(note = "stdprot")
}
\author{
Lang Zhou
}
