% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_upse.R
\name{plot_upse}
\alias{plot_upse}
\title{plot UltraPseR result}
\usage{
plot_upse(seq_result, line_size = 1, alpha = 0.7, spline_shape = 0.4)
}
\arguments{
\item{seq_result}{Data.frame, the result after running UpseR()}

\item{line_size}{line size}

\item{alpha}{alpha}

\item{spline_shape}{A numeric vector of values between -1 and 1, which control the shape of the spline relative to the control points.}
}
\value{
ggplot2 object
}
\description{
plot UltraPseR result
}
\examples{
tiny_sequences <- system.file("extdata", "tiny.fas", package = "UltraPseR")
y <- UpseR(input = tiny_sequences, 
           note = "stdprot", module = "pse", property = c("ZIMJ680105"), 
           w = 0.1, l = 10, t = 1)
}
\author{
Lang Zhou
}
