% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpseR.R
\name{UpseR}
\alias{UpseR}
\title{wappers UltraPse in R}
\usage{
UpseR(
  input,
  format = "svm",
  module = "comp",
  note = "stdprot",
  property = "ZIMJ680105",
  w = 0.05,
  l = 10,
  t = 1
)
}
\arguments{
\item{input}{A FASTA format file for input. The comment line of a sequence in this file should be unique. If you do not specify an in-file, the program will try to read it from keyboard.}

\item{format}{Output format.{svm|tsv|csv}}

\item{module}{Representation module choices. This option can appear for multiple times.{comp|pse|dpc|cov|lua|pseb3|user,LIB_OBJECT,MOD_NAME}}

\item{note}{Sequence definition types. There are four internal types. If the user want to use their own sequence type, a task definition file id required. {stdprot|didna|dirna|tridna})}

\item{property}{Activate the property PROP-ID for current task. This option can appear for multiple times.}

\item{w}{Omega,a mode specific option. Omega parameter in the pse module. This option can have user defined purpose.}

\item{l}{Lambda,a mode specific option. Lambda parameter in the pse module. This option can have user defined purpose.}

\item{t}{Type, a mode specific option. Type parameter in the pse / cov module. This option can have user defined purpose.}
}
\description{
wappers UltraPse in R
}
\examples{
tiny_sequences <- system.file("extdata", "tiny.fas", package = "UltraPseR")
UpseR(input = tiny_sequences, format = "svm", note = "stdprot", module = "pse", property = "ZIMJ680105", w = 0.05, l = 10, t = 2)
}
\author{
Lang Zhou
}
